src/reporting_system/requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
pydantic==2.12.3
python-dotenv==1.2.1
celery==5.5.3
redis 
gevent 
aioodbc==0.5.0
pyodbc==5.2.0
psycopg2-binary==2.9.10
crewai
langchain-openai
langgraph
numpy
pandas
sentence-transformers
SQLAlchemy
litellm